Renamed the C++ header (.h -> .hpp)
